completed refactoring of add-account, next is create
