+++ src/components/add-account.vue
@@ -93,119 +93,23 @@
             </div>
         </div>
         <div
-            v-else-if="step==2 && (selectedChain!='BTS' || BTSImportType=='1')"
+            v-else-if="step==2"
             id="step2"
         >
-            <h4 class="h4 mt-3 font-weight-bold">
-                {{ $t('step_counter',{ 'step_no' : 2}) }}
-            </h4>
-            <p
-                v-if="accessType=='account'"
-                class="mb-2 font-weight-bold"
-            >
-                {{ $t('account_name',{ 'chain' : selectedChain}) }}
-            </p>
-            <p
-                v-else
-                class="mb-2 font-weight-bold"
-            >
-                {{ $t('address_name',{ 'chain' : selectedChain}) }}
-            </p>
-            <input
-                id="inputAccount"
-                v-model="accountname"
-                type="text"
-                class="form-control mb-3"
-                :placeholder="$t('account_name',{ 'chain' : selectedChain})"
-                required
-            >
-            <p class="my-3 font-weight-normal">
-                {{ $t('keys_cta') }}
-            </p>
-            <template v-if="requiredFields.active !== null">
-                <p
-                    v-if="accessType=='account'"
-                    class="mb-2 font-weight-bold"
-                >
-                    {{ $t('active_authority') }}
-                </p>
-                <p
-                    v-else
-                    class="mb-2 font-weight-bold"
-                >
-                    {{ $t('public_authority') }}
-                </p>
-                <input
-                    id="inputActive"
-                    v-model="activepk"
-                    type="password"
-                    class="form-control mb-3 small"
-                    :placeholder="accessType=='account' ? $t('active_authority_placeholder') : $t('public_authority_placeholder')"
-                    required
-                >
-            </template>
-            <template v-if="requiredFields.memo !== null">
-                <p class="mb-2 font-weight-bold">
-                    {{ $t('memo_authority') }}
-                </p>
-                <input
-                    id="inputMemo"
-                    v-model="memopk"
-                    type="password"
-                    class="form-control mb-3 small"
-                    :placeholder="$t('memo_authority_placeholder')"
-                    required
-                >
-            </template>
-            <template v-if="requiredFields.owner !== null">
-                <b-form-checkbox
-                    id="incOwnerCB"
-                    v-model="includeOwner"
-                    value="1"
-                    unchecked-value="0"
-                    class="mb-3"
-                >
-                    {{ $t('include_owner_check') }}
-                </b-form-checkbox>
-                <div v-if="includeOwner==1">
-                    <input
-                        id="inputOwner"
-                        v-model="ownerpk"
-                        type="password"
-                        class="form-control mb-3 small"
-                        :placeholder="$t('owner_authority_placeholder')"
-                        required
-                    >
-                </div>
-            </template>
-            <div class="row">
-                <div class="col-6">
-                    <button
-                        class="btn btn-lg btn-primary btn-block"
-                        type="submit"
-                        @click="step1"
-                    >
-                        {{ $t('back_btn') }}
-                    </button>
-                </div>
-                <div class="col-6">
-                    <button
-                        class="btn btn-lg btn-primary btn-block"
-                        type="submit"
-                        @click="step3"
-                    >
-                        {{ $t('next_btn') }}
-                    </button>
-                </div>
-            </div>
-        </div>
-        <div
-            v-if="step==2 && selectedChain=='BTS' && BTSImportType=='2'"
-            id="step2"
-        >
+            <ImportKeys
+                v-if="selectedChain != 'BTS' || BTSImportType=='1'"
+                ref="import_bitshares_accounts"
+                :selectedChain="selectedChain"
+            />
             <ImportCloudPass
-                    ref="import_bitshares_cloud"
-                    :selectedChain="selectedChain"
+                v-else-if="selectedChain=='BTS' && BTSImportType=='2'"
+                ref="import_bitshares_accounts"
+                :selectedChain="selectedChain"
+            />
+            <ImportBinFile
+                v-else="selectedChain=='BTS' && BTSImportType=='3'"
+                ref="import_bitshares_accounts"
+                :selectedChain="selectedChain"
             />
             <div class="row">
                 <div class="col-6">
@@ -229,153 +133,6 @@
             </div>
         </div>
         <div
-            v-if="step==2 && selectedChain=='BTS' && BTSImportType=='3'"
-            id="step2"
-        >
-            <h4 class="h4 mt-3 font-weight-bold">
-                {{ $t('step_counter',{ 'step_no' : 2}) }}
-            </h4>
-            <template v-if="substep1">
-                <p                
-                    class="mb-2 font-weight-bold"
-                >
-                    {{ $t('Select your .bin backup file.') }}
-                </p>
-                <input
-                    v-if="step1"
-                    type="file"
-                    class="form-control mb-3 small" 
-                    @change="handleWalletSelect"
-                >
-                <p                
-                    class="mb-2 font-weight-bold"
-                >
-                    {{ $t('Enter your .bin file password.') }}
-                </p>
-                <input
-                    v-model="wallet_pass"
-                    type="password"
-                    class="form-control mb-3 small"
-                >
-                <div class="row">
-                    <div class="col-6">
-                        <button
-                            class="btn btn-lg btn-primary btn-block"
-                            type="submit"
-                            @click="step1"
-                        >
-                            {{ $t('back_btn') }}
-                        </button>
-                    </div>
-                    <div class="col-6">
-                        <button
-                            class="btn btn-lg btn-primary btn-block"
-                            type="submit"
-                            @click="decryptBackup"
-                        >
-                            {{ $t('next_btn') }}
-                        </button>
-                    </div>
-                </div>
-            </template>
-            <template v-if="substep2">
-                <div class="import-accounts mt-3">
-                    <table class="table small table-striped table-sm">
-                        <thead>
-                            <tr>                    
-                                <th
-                                    rowspan="2"
-                                    class="align-middle"
-                                >
-                                    Account Name
-                                </th>
-                                <th
-                                    colspan="2"
-                                    class="align-middle"
-                                >
-                                    Active Authority
-                                </th>
-                                <th
-                                    colspan="2"
-                                    class="align-middle"
-                                >
-                                    Owner Authority
-                                </th>
-                                <th
-                                    rowspan="2"
-                                    class="align-middle"
-                                >
-                                    Memo
-                                </th>
-                                <th
-                                    rowspan="2"
-                                    class="align-middle"
-                                >
-                                    Import?
-                                </th>
-                            </tr>                
-                            <tr>
-                                <th class="align-middle">
-                                    Propose
-                                </th>
-                                <th class="align-middle">
-                                    Transact
-                                </th>
-                                <th class="align-middle">
-                                    Propose
-                                </th>
-                                <th class="align-middle">
-                                    Transact
-                                </th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <tr
-                                v-for="account in accounts" 
-                                :key="account.id"
-                                :class="{ disabledImport : !account.importable}"
-                            >                    
-                                <td class="text-center align-middle">
-                                    {{ account.name }}<br>({{ account.id }})
-                                </td>
-                                <td class="text-center align-middle">
-                                    {{ account.active.canPropose ? 'Y' : 'N' }}
-                                </td>
-                                <td class="text-center align-middle">
-                                    {{ account.active.canTransact ? 'Y' : 'N' }}
-                                </td>
-                                <td class="text-center align-middle">
-                                    {{ account.owner.canPropose ? 'Y' : 'N' }}
-                                </td>
-                                <td class="text-center align-middle">
-                                    {{ account.owner.canTransact ? 'Y' : 'N' }}
-                                </td>
-                                <td class="text-center align-middle">
-                                    {{ account.memo.canSend ? 'Y' : 'N' }}
-                                </td>
-                                <td class="text-center align-middle">
-                                    <input
-                                        v-if="account.importable"
-                                        :id="account.name"
-                                        v-model="picked"
-                                        type="checkbox"
-                                        :value="account.id"
-                                    >
-                                </td>
-                            </tr>
-                        </tbody>
-                    </table>
-                </div>
-                <button
-                    v-if="picked.length>0"
-                    class="btn btn-lg btn-primary btn-block mt-3"
-                    @click="simpleStep3"
-                >
-                    {{ $t('import_btn') }}
-                </button>
-            </template>
-        </div>
-        <div
             v-if="step==3"
             id="step3"
         >
@@ -393,20 +150,10 @@
                 :placeholder="$t('password_placeholder')"
                 required
             >
-      
             <button
-                v-if="BTSImportType!='3'"
                 class="btn btn-lg btn-primary btn-block"
                 type="submit"
-                @click="verifyAndCreate"
-            >
-                {{ $t('next_btn') }}
-            </button>
-            <button
-                v-if="BTSImportType=='3'"
-                class="btn btn-lg btn-primary btn-block"
-                type="submit"
-                @click="importAccounts"
+                @click="addAccounts"
             >
                 {{ $t('next_btn') }}
             </button>
@@ -430,40 +177,31 @@
 <script>
     import { blockchains } from "../config/config.js";
     import getBlockchain from "../lib/blockchains/blockchainFactory";
-    import { EventBus } from "../lib/event-bus.js";
-    import RendererLogger from "../lib/RendererLogger";    
-    import {PrivateKey} from "bitsharesjs";
-    import BTSWalletHandler from "../lib/blockchains/bitshares/BTSWalletHandler";
-    import ImportCloudPass from "./blockchains/bitshares/ImportBinFile";
 
+    import ImportCloudPass from "./blockchains/bitshares/ImportCloudPass";
+    import ImportBinFile from "./blockchains/bitshares/ImportBinFile";
+    import ImportKeys from "./blockchains/ImportKeys";
+
+    import { EventBus } from "../lib/event-bus.js";
+
+    import RendererLogger from "../lib/RendererLogger";
     const logger = new RendererLogger();
 
     export default {
         name: "AddAccount",
-        components: {ImportCloudPass},
+        components: {ImportKeys, ImportCloudPass, ImportBinFile},
         i18nOptions: { namespaces: "common" },
         data() {
             return {
                 accountname: "",
-                accountID: "",
-                activepk: "",
-                ownerpk: "",
-                memopk: "",
                 password: "",
                 step: 1,
-                substep1: true,
-                substep2:false,
                 s1c: "",
-                bitshares_cloud_login_password: "",
                 includeOwner: 0,
                 errorMsg: "",
                 selectedChain: 0,
                 BTSImportType: 0,
                 chainList: Object.values(blockchains),
-                wallet_file: null,
-                wallet_pass: null,
-                accounts:[],
-                picked:[]
             };
         },
         mounted() {
@@ -474,180 +212,70 @@
                 this.step = 1;
             },
             step2: function() {
-                if (this.selectedChain!=0) {
-                    if (this.selectedChain!='BTS' || this.BTSImportType!=0) {
-                        let blockchain = getBlockchain(this.selectedChain);
-                        this.accessType = blockchain.getAccessType();
-                        this.requiredFields = blockchain.getSignUpInput();
-                        this.step = 2;
-                    }
-                }
+                this.step = 2;
             },
             step3: async function() {
-                if (this.accountname == "") {
-                    this.errorMsg = this.$t("missing_account_error", {
-                        chain: this.selectedChain
-                    });
-                    this.$refs.errorModal.show();
-                    return;
-                }
-
                 EventBus.$emit("popup", "load-start");
                 try {
                     let blockchain = getBlockchain(this.selectedChain);
                     // abstract UI concept more
-                    let authorities = null;
-                    let account=null;
                     if (blockchain.getAccessType() == "account") {
-                        if (this.BTSImportType==2){
-                            let account = this.$refs.import_bitshares_cloud.getAccountEvent();
-
-                            this.password = account.password;
-                            this.accountname = account.accountName;
-                            this.accountID = account.accountID;
-                            this.activepk = account.keys.active;
-                            this.ownerpk = account.keys.owner;
-                            this.memopk = account.keys.active;
-                        }else{
-                            authorities = {
-                                active: this.activepk,
-                                memo: this.memopk,
-                                owner: this.includeOwner == 1 ? this.ownerpk : null
-                            };
-                            account = await blockchain.verifyAccount(this.accountname, authorities);
-                        }
+                        this.accounts_to_import = await this.$refs.import_bitshares_bin.getAccountEvent();
                     } else {
-                        authorities = {
+                        let authorities = {
                             active: this.activepk
                         };
-                        account = await blockchain.verifyAccount(this.accountname, authorities);
+                        let account = await blockchain.verifyAccount(this.accountname, authorities);
+                        this.accounts_to_import = [ account ];
                     }
-                    
                     EventBus.$emit("popup", "load-end");
-                    this.accountID = account.id;
-                    this.step = 3;
-                } catch (err) {
-                    this.accountID = "";
-                    if (err.key) {
-                        this.errorMsg = this.$t(err.key);
-                    } else {
-                        this.errorMsg = err.toString();
+                    if (this.accounts_to_import != null) {
+                        // if import accounts are filled, advance to next step. If not, it is a substep in the
+                        // import component
+                        this.step = 3;
                     }
-                    this.$refs.errorModal.show();
+                } catch (err) {
+                    this._handleError(err);
                 } finally {
                     EventBus.$emit("popup", "load-end");
                 }
             },
-            simpleStep3: function () {
-                this.step=3;
-            },
-            getAuthoritiesFromPass: function(password,legacy=false) {
-                let active_seed = this.accountname + 'active' + password;
-                let owner_seed = this.accountname + 'owner' + password;
-                let memo_seed = this.accountname + 'memo' + password;
-                if (legacy) {
-                    return {
-                        active: PrivateKey.fromSeed(active_seed).toWif(),
-                        memo: PrivateKey.fromSeed(active_seed).toWif(),
-                        owner: PrivateKey.fromSeed(owner_seed).toWif()
-                    };
-                }else{
-                    return {
-                        active: PrivateKey.fromSeed(active_seed).toWif(),
-                        memo: PrivateKey.fromSeed(memo_seed).toWif(),
-                        owner: PrivateKey.fromSeed(owner_seed).toWif()
-                    };
+            _handleError(err) {
+                if (err == 'invalid') {
+                    this.errorMsg = this.$t("invalid_password");
+                } else if (err == 'update_failed') {
+                    this.errorMsg = this.$t("update_failed");
+                } else if (err.key) {
+                    this.errorMsg = this.$t(err.key);
+                } else {
+                    this.errorMsg = err.toString();
                 }
-            },            
-            handleWalletSelect: function(e) {
-                this.wallet_file=e.target.files[0];
+                this.$refs.errorModal.show();
             },
-            decryptBackup: function() {
-                EventBus.$emit("popup", "load-start");
-                let reader = new FileReader();
-                reader.onload = async evt => {
-                    let wh=new BTSWalletHandler(evt.target.result);
-                    try {
-                        let unlocked=await wh.unlock(this.wallet_pass);
-                        if (unlocked) {
-                            this.accounts=await wh.lookupAccounts();
-                            console.log(this.accounts);
-                            this.substep1=false;
-                            this.substep2=true;
-                            EventBus.$emit("popup", "load-end");
-                        }
-                    }catch(err) {
-                        if (err.key) {
-                            this.errorMsg = this.$t(err.key);
-                        } else {
-                            this.errorMsg = err.toString();
-                        }
-                        this.$refs.errorModal.show();
-                    }finally {
-                        EventBus.$emit("popup", "load-end");
-                    }
-                };
-                reader.readAsBinaryString(this.wallet_file);
-            },
-            importAccounts: async function() {
-                let toimport=this.accounts.filter(x => this.picked.includes(x.id));
-                for (let i in toimport) {
-                    let account=toimport[i];
-                    await  this.$store
-                        .dispatch("AccountStore/addAccount", {
-                            password: this.password,
-                            account: {
-                                accountName: account.name,
-                                accountID: account.id,
-                                chain: this.selectedChain,
-                                keys: {
-                                    active: account.active.key,
-                                    owner: account.owner.key,
-                                    memo: account.memo.key
-                                }
-                            }
-                        });
-
-                }
-                this.$router.replace("/dashboard");
-            },
-            verifyAndCreate: async function() {
+            addAccounts: async function() {
                 if (this.password == "") {
                     this.$refs.errorModal.show();
                     this.errorMsg = this.$t("empty_pass_error");
                     return;
                 }
-
                 EventBus.$emit("popup", "load-start");
+                try {
+                    if (!!this.accounts_to_import) {
+                        for (let i in this.accounts_to_import) {
+                            let account = this.accounts_to_import[i];
+                            account.password = this.password;
+                            account.walletname = this.walletname;
+                            await this.$store.dispatch("AccountStore/addAccount", account);
 
-                if (this.accountID !== null) {
-                    this.$store
-                        .dispatch("AccountStore/addAccount", {
-                            password: this.password,
-                            account: {
-                                accountName: this.accountname,
-                                accountID: this.accountID,
-                                chain: this.selectedChain,
-                                keys: {
-                                    active: this.activepk,
-                                    owner: this.ownerpk,
-                                    memo: this.memopk
-                                }
-                            }
-                        })
-                        .then(() => {
-                            this.$router.replace("/dashboard");
-                        }).catch((e) => {
-                            EventBus.$emit("popup", "load-end");
-                            if (e=='invalid') {
-                                this.$refs.errorModal.show();
-                                this.errorMsg = this.$t("invalid_password");
-                            }
-                            if (e=='update_failed') {
-                                this.$refs.errorModal.show();
-                                this.errorMsg = this.$t("update_failed");
-                            }
-                        });
+                        }
+                        this.$router.replace("/dashboard");
+                    } else {
+                        throw "No account selected!";
+                    }
+                } catch (err) {
+                    this._handleError(err);
+                } finally {
+                    EventBus.$emit("popup", "load-end");
                 }
             }
         }
